% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSfilesPath.R
\name{getSfilesPath}
\alias{getSfilesPath}
\title{Collecting paths of some specified files that you want to import or read.}
\usage{
getSfilesPath(root.path, filename)
}
\arguments{
\item{root.path}{is the root path including all folders and files that you would like to search.}

\item{filename}{is the name of files that you want to collect.}
}
\value{
The whole paths of all files that meet the criteria were saved as a list.
}
\description{
If you want to collect all files under certain folder, this function should be the perfect one. It will collect all files with certain name. Then this function will return a list will all paths of those files so that further import or read is feasible.
}
\examples{
getSfilesPath(root.path = R.home(), filename = "?.exe")
}

