% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{MergerXLSX is a function to merger of xlsx files by identified variable.}
\usage{
hello(name)
}
\arguments{
\item{name}{is a text variable of function user.}
}
\value{
Return data are all original data with addin variables.
}
\description{
This is a test function.
}
\details{
This function is used to test.
}
\author{
Xiaorui.Zhu
}

