% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltodata.R
\name{htmltodata}
\alias{htmltodata}
\title{htmltodata}
\usage{
htmltodata(path)
}
\arguments{
\item{path}{is the path of the file that you want to import into R and then export.}
}
\value{
The return data are a list include all text results of submitters' answers.
}
\description{
"htmltodata" function is used to transfer information from html files to R or xlsx files
}
\author{
Xiaorui(Jeremy) Zhu
}

