% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolida.R
\name{consolida}
\alias{consolida}
\title{consolida funtion is an internal function for data merger.}
\usage{
consolida(row, data, mergeVar)
}
\arguments{
\item{row}{is each sample that is already divided from the original file.}

\item{data}{is the "addin" file.}

\item{mergeVar}{is the variable that use to merge.}
}
\value{
is single line contains original variables and addin variables.
}
\description{
This is a function that use to match original data and addin data with identified variable.
}
\details{
This function is for internal use only, so no need to export it. It figures out the ID in the "addin" file then merge variables in addin file to the original file. This function is used for further "lapply" porcess.
}
\author{
Xiaorui.Zhu
}

