% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergerXLSX.R
\name{MergerXLSX}
\alias{MergerXLSX}
\title{MergerXLSX is a function to merger of xlsx files by identified variable.}
\usage{
MergerXLSX(original_file, addin_file, mergeID)
}
\arguments{
\item{original_file}{is the name of original file. This file contains all original data. This file should be "xlsx" file and save in the same folder.}

\item{addin_file}{is the file that need to be merged. This file should be "xlsx" file and save in the same folder.}

\item{mergeID}{is the merger variable name in both files.}
}
\value{
Return data are all original data with addin variables.
}
\description{
This is a function that can be used to merger xlsx file using identified variables.
}
\details{
This function need three parameters. First is the name of the file with which need to be merged. Second is the name of original file that contains data. Third is the identifiable variable name that in both files.
}
\author{
Xiaorui.Zhu
}

